+++ src/main/java/subway/SubwayApplication.java
@@ -6,8 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class SubwayApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(SubwayApplication.class, args);
-	}
-
+    public static void main(String[] args) {
+        SpringApplication.run(SubwayApplication.class, args);
+    }
 }
