+++ src/main/java/subway/ui/SectionController.java
@@ -2,9 +2,9 @@ package subway.ui;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.application.SectionService;
 import subway.dto.LineStationRequest;
 import subway.dto.LineStationResponse;
+import subway.service.SectionService;
 
 import javax.validation.Valid;
 import java.net.URI;
