+++ src/main/java/subway/ui/StationController.java
@@ -2,9 +2,9 @@ package subway.ui;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.application.StationService;
 import subway.dto.StationRequest;
 import subway.dto.StationResponse;
+import subway.service.StationService;
 
 import javax.validation.Valid;
 import java.net.URI;
