+++ src/test/java/subway/domain/SectionsRemoveTest.java
@@ -1,6 +1,5 @@
 package subway.domain;
 
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
